# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Esta é a API para o projeto de banco de dados, permitindo a gestão de alunos, cursos e o registro de interesse em turmas.",
        "title": "API do Sistema de Interesse de Matrícula - UnB",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/alunos": {
            "get": {
                "description": "Retorna uma lista com todos os alunos cadastrados no banco de dados",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Lista todos os alunos",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Aluno"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Cria um novo aluno com base nos dados enviados em JSON",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Cria um novo aluno",
                "parameters": [
                    {
                        "description": "Dados do Aluno para Criar",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Aluno"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/alunos/{matricula}": {
            "get": {
                "description": "Retorna os detalhes de um único aluno com base na sua matrícula",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Busca um aluno pela matrícula",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matrícula do Aluno",
                        "name": "matricula",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Atualiza os dados de um aluno com base na sua matrícula",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Atualiza um aluno existente",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matrícula do Aluno",
                        "name": "matricula",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dados do Aluno para Atualizar",
                        "name": "aluno",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.Aluno"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Aluno"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Apaga um aluno do banco de dados com base na sua matrícula",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Apaga um aluno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matrícula do Aluno",
                        "name": "matricula",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/alunos/{matricula}/foto": {
            "post": {
                "description": "Recebe um arquivo de imagem e o salva para o aluno com a matrícula especificada",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Alunos"
                ],
                "summary": "Faz o upload de uma foto para um aluno",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Matrícula do Aluno",
                        "name": "matricula",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Arquivo de foto para upload",
                        "name": "foto",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Aluno": {
            "type": "object",
            "properties": {
                "cod_curso": {
                    "type": "integer"
                },
                "cpf": {
                    "type": "string"
                },
                "data_nascimento": {
                    "type": "string"
                },
                "email_institucional": {
                    "type": "string"
                },
                "email_pessoal": {
                    "type": "string"
                },
                "matricula": {
                    "type": "string"
                },
                "nacionalidade": {
                    "type": "string"
                },
                "nome_completo": {
                    "type": "string"
                },
                "semestre_ingresso": {
                    "type": "string"
                }
            }
        },
...
    }
}